+++ modules/Grades/GradebookBreakdown.php
@@ -80,7 +80,8 @@ $assignments_RET = DBGet( "SELECT ASSIGNMENT_ID,TITLE,POINTS
 	AND MARKING_PERIOD_ID='" . UserMP() . "'
 	ORDER BY " . DBEscapeIdentifier( Preferences( 'ASSIGNMENT_SORTING', 'Gradebook' ) ) . " DESC" );
 
-$assignment_select = '<select name="assignment_id" id="assignment_id" onchange="ajaxPostForm(this.form)">';
+// @since 12.5 CSP remove unsafe-inline Javascript
+$assignment_select = '<select name="assignment_id" id="assignment_id" class="onchange-ajax-post-form">';
 
 $assignment_select .= '<option value="totals"' . ( $_REQUEST['assignment_id'] === 'totals' ? ' selected' : '' ) . '>' .
 _( 'Totals' ) .