+++ modules/Users/TeacherPrograms.php
@@ -75,8 +75,9 @@ if ( UserStaffID() )
 		SetUserCoursePeriod( $_REQUEST['period'] );
 	}
 
+	// @since 12.5 CSP remove unsafe-inline Javascript
 	$period_select = '<label for="period" class="a11y-hidden">' . _( 'Course Periods' ) . '</label>
-		<select name="period" id="period" autocomplete="off" onChange="ajaxPostForm(this.form);">';
+		<select name="period" id="period" autocomplete="off" class="onchange-ajax-post-form">';
 
 	$optgroup = $current_cp_found = false;