+++ CHANGES.md
@@ -86,6 +86,7 @@ Changes in 12.5
 - CSP add .onchange-maybe-edit-select CSS class in StudentsUsersInfo.fnc.php & Address.inc.php
 - Prevent using username, or email in the password in General_Info.inc.php
 - CSP add .onclick-ajax-link & .onchange-ajax-link CSS classes, use `data-link` attribute, program wide
+- CSP add .onchange-ajax-post-form CSS class, program wide
 
 Changes in 12.4.3
 -----------------